Refactor CustomCalendar Component: Extract Logic into Custom Hook
Performed a comprehensive refactor of the `CustomCalendar` React Native component, where a portion of its logic has been externalized to a custom hook named `useCustomCalendar`. Here are the detailed changes:

1.Created a Custom Hook:
   - A custom hook `useCustomCalendar` encapsulating logic related to handling selected dates and modal visibility has been created.
   - The hook leverages `useState` and `useCallback` to manage state and returns several values to handle date selection, mark unavailable dates, and handle modal actions.

2.Modified the CustomCalendar Component:
   - The original component contained all logic, whereas in the modified version, the logic has been extracted into the custom hook.
   - Removed import of `useState` and `useCallback`, added import of `useCustomCalendar`, and extracted its returned values within the component.
   - Logic related to date handling and modal visibility was moved into the custom hook.

3.implification and Separation of Responsibilities:
   - The refactor has simplified the `CustomCalendar` component, segregating the logic related to date selection and modal manipulation into an independent hook.
   - This enhances understandability, maintainability, and testing, while also allowing for potential reusability of the hook elsewhere.

This change emphasizes best practices by fostering a cleaner, more modular design.
